pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    { name = "", email = "" },
]
dependencies = [
    "six>=1.16.0",
    "requests>=2.30.0",
    "redis>=4.5.4",
    "bs4>=0.0.1",
    "pillow>=9.5.0",
    "numpy>=1.24.3",
    "openpyxl>=3.1.2",
    "ruamel-yaml>=0.17.22",
]
requires-python = ">=3.10"
license = { text = "MIT" }
